requirements.txt
aiogram==2.25.1
openai==0.27.8
replicate==0.8.0
python-dotenv==1.0.0